add support for mingw32 compiler
i686-w64-mingw32-gcc compiler is used.
to build cross-compile svm for Windows under GNU/Linux use
following command:
$ ./adm/makewinvm.py -c mingw32

+++ adm/makewinvm.py
@@ -9,14 +9,19 @@
 #
 
 import os
+import sys
 import env
 import argparse
 import fileutil
 import compilewin
 import compilekit
 import platArchive
+import xml.dom.minidom
 
-platFile = os.path.join(env.platforms, "src", "generic", "win32", "generic-win32.xml")
+compiler = "msvc"
+# compiler = "mingw32"
+
+platFile = ""
 
 exeFile = os.path.join(env.bin, "svm.exe")
 
@@ -36,6 +41,21 @@ defs = { 'WIN32':None, '_WIN32':None, 'SOCKET_FAMILY_INET':None,
          'PLAT_BUILD_VERSION':'\\"' + env.buildVersion() + '\\"' }
 
 
+def platformType(string):
+  if xml.dom.minidom.parse(string).documentElement.nodeName != "sedonaPlatform":
+    msg = "" + string + " is not a sedonaPlatform XML file"
+    raise argparse.ArgumentTypeError(msg)
+  global platFile
+  platFile = string
+
+
+def compilerType(string):
+  if not getattr(compilewin, string, None):
+    msg =  "Compiler '" + string + "' is not supported."
+    raise argparse.ArgumentTypeError(msg)
+  global compiler
+  compiler = string
+
 # initParser
 def initParser():
   global parser
@@ -45,26 +65,55 @@ def initParser():
                              help='Set SVM version string to VERSION', 
                              metavar="VERSION")
 
+  parser.add_argument('-l', '--list-compilers', action='store_true',
+                             help='List the supported compilers')
+
+  parser.add_argument('-p', '--platform', type=platformType, action='store',
+                             help='The path to the sedonaPlatform XML file. \nEnvironment variable SVM_PLATFORM can be used to set the platform path.',
+                             metavar="SVM_PLATFORM")
+  parser.add_argument('-c', '--compiler', type=compilerType, action='store',
+                             help="The compiler to use. Defaults to 'msvc'", metavar="COMPILER")
+
   ipgroup = parser.add_mutually_exclusive_group()
   ipgroup.add_argument('-4', '--ipv4', action='store_true', default=False,  
                              help='Use IPv4 protocol (default)')
   ipgroup.add_argument('-6', '--ipv6', action='store_true', default=False, 
                              help='Use IPv6 protocol')
 
+# list compilers
+def listCompilers():
+  compilers = [e for e in dir(compilewin) if callable(getattr(compilewin,e))]# and not e.startsWith("__")]
+  if len(compilers) == 0:
+    print "No compilers supported for building the SVM"
+  else:
+    print "Compiler(s) supported for building the SVM:"
+    for c in compilers:
+      print " " + c
+
+
 # compile
 def compile(cdefs=defs):
   try:
+    if xml.dom.minidom.parse(platFile).documentElement.nodeName != "sedonaPlatform":
+      raise env.BuildError("" + platFile + " is not a sedonaPlatform XML file")
+
     fileutil.rmdir(stageDir)
     compilekit.compile(platFile + " -outDir " + stageDir)
-    #compilewin.compile(exeFile, srcFiles, includes, libs, cdefs, ["/P"])  # preproc output
-    compilewin.compile(exeFile, srcFiles, includes, libs, cdefs)
+    getattr(compilewin, compiler)(exeFile, srcFiles, includes, libs, cdefs)
 
   except env.BuildError:
     print "**"
     print "** FAILED [" + exeFile + "]"
     print "**"
-    return 1
+    sys.exit(1)
 
+
+def verifyOpts():
+  global platFile
+  if not platFile:
+    platFile = os.environ.get("SVM_PLATFORM")
+    if not platFile:
+      platFile = os.path.join(env.platforms, "src", "generic", "win32", "generic-win32.xml")
 
 
 # Main
@@ -75,6 +124,16 @@ if __name__ == '__main__':
   # Parse command line arguments
   initParser()
   options = parser.parse_args()
+
+  if (options.list_compilers):
+    listCompilers()
+    sys.exit(0)
+
+  if options.compiler:
+    compiler = options.compiler
+
+  if options.platform:
+    platFile = options.platform
 
   # Add command line arg to select ipv4 vs. ipv6 socket family
   if (options.ipv6):
@@ -90,6 +149,8 @@ if __name__ == '__main__':
   # Add cmd line arg to set version string
   if options.ver:
     defs["PLAT_BUILD_VERSION"] = '\\"' + options.ver + '\\"'
+
+  verifyOpts()
 
   # Compile Sedona VM
   compile(config)
